requirements.txt
tensorflow==2.13.0
streamlit==1.37.0
pillow
numpy
